K-means: switch pipeline to 16 bit.
Less RAM, less for loops in SIMD, simpler SIMD instructions.

Speed-ups: being computed.

+++ av1/encoder/k_means_template.h
@@ -28,7 +28,7 @@
 // Though we want to compute the smallest L2 norm, in 1 dimension,
 // it is equivalent to find the smallest L1 norm and then square it.
 // This is preferrable for speed, especially on the SIMD side.
-static int RENAME(calc_dist)(const int *p1, const int *p2) {
+static int RENAME(calc_dist)(const int16_t *p1, const int16_t *p2) {
 #if AV1_K_MEANS_DIM == 1
   return abs(p1[0] - p2[0]);
 #else
@@ -41,7 +41,7 @@ static int RENAME(calc_dist)(const int *p1, const int *p2) {
 #endif
 }
 
-void RENAME(av1_calc_indices)(const int *data, const int *centroids,
+void RENAME(av1_calc_indices)(const int16_t *data, const int16_t *centroids,
                               uint8_t *indices, int64_t *dist, int n, int k) {
   if (dist) {
     *dist = 0;
@@ -67,7 +67,7 @@ void RENAME(av1_calc_indices)(const int *data, const int *centroids,
   }
 }
 
-static void RENAME(calc_centroids)(const int *data, int *centroids,
+static void RENAME(calc_centroids)(const int16_t *data, int16_t *centroids,
                                    const uint8_t *indices, int n, int k) {
   int i, j;
   int count[PALETTE_MAX_SIZE] = { 0 };
@@ -98,11 +98,11 @@ static void RENAME(calc_centroids)(const int *data, int *centroids,
   }
 }
 
-void RENAME(av1_k_means)(const int *data, int *centroids, uint8_t *indices,
-                         int n, int k, int max_itr) {
-  int centroids_tmp[AV1_K_MEANS_DIM * PALETTE_MAX_SIZE];
+void RENAME(av1_k_means)(const int16_t *data, int16_t *centroids,
+                         uint8_t *indices, int n, int k, int max_itr) {
+  int16_t centroids_tmp[AV1_K_MEANS_DIM * PALETTE_MAX_SIZE];
   uint8_t indices_tmp[MAX_PALETTE_BLOCK_WIDTH * MAX_PALETTE_BLOCK_HEIGHT];
-  int *meta_centroids[2] = { centroids, centroids_tmp };
+  int16_t *meta_centroids[2] = { centroids, centroids_tmp };
   uint8_t *meta_indices[2] = { indices, indices_tmp };
   int i, l = 0, prev_l, best_l = 0;
   int64_t this_dist;
